# Drop JDK 8 support for more efficient parsing and serialization of some types of numbers and `java.time._` classes