--- a/.github/workflows/build.yml
+++ b/.github/workflows/build.yml
@@ -21,7 +21,7 @@ jobs:
       matrix:
         java:
           - zulu@1.11
-          - zulu@1.8
+          - zulu@1.17
         os:
           - ubuntu-latest
           - macOS-latest
@@ -44,7 +44,7 @@ jobs:
         run: "sbt -batch clean +test +mimaReportBinaryIssues"
         shell: bash
       - name: Test with coverage
-        if: ${{ matrix.os == 'ubuntu-latest' && matrix.java == 'zulu@1.8' }}
+        if: ${{ matrix.os == 'ubuntu-latest' && matrix.java == 'zulu@1.11' }}
         env:
           COVERALLS_REPO_TOKEN: ${{ secrets.COVERALLS_REPO_TOKEN }}
         run: "sbt -batch ++2.13.6! coverage jsoniter-scala-coreJVM/test jsoniter-scala-macrosJVM/test jsoniter-scala-benchmarkJVM/test coverageAggregate coveralls"