--- a/jsoniter-scala-core/jvm/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonReader.scala
+++ b/jsoniter-scala-core/jvm/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonReader.scala
@@ -1415,7 +1415,7 @@ final class JsonReader private[jsoniter_scala](
     else if (e >= 310) Double.PositiveInfinity
     else {
       var shift = java.lang.Long.numberOfLeadingZeros(m)
-      var mant = multiplyHigh(m << shift, pow10Mantissas(e.toInt + 343)) // FIXME: Use Math.multiplyHigh after dropping JDK 8 support
+      var mant = unsignedMultiplyHigh(m << shift, pow10Mantissas(e.toInt + 343)) // FIXME: replace by Math.unsignedMultiplyHigh after dropping of JDK 17 support
       var exp = addExp(-shift, e.toInt)
       shift = java.lang.Long.numberOfLeadingZeros(mant)
       mant <<= shift
@@ -1554,7 +1554,7 @@ final class JsonReader private[jsoniter_scala](
     else if (e >= 39) Float.PositiveInfinity
     else {
       var shift = java.lang.Long.numberOfLeadingZeros(m)
-      var mant = multiplyHigh(m << shift, pow10Mantissas(e.toInt + 343)) // FIXME: Use Math.multiplyHigh after dropping JDK 8 support
+      var mant = unsignedMultiplyHigh(m << shift, pow10Mantissas(e.toInt + 343)) // FIXME: replace by Math.unsignedMultiplyHigh after dropping of JDK 17 support
       var exp = addExp(-shift, e.toInt)
       shift = java.lang.Long.numberOfLeadingZeros(mant)
       mant <<= shift
@@ -1586,17 +1586,8 @@ final class JsonReader private[jsoniter_scala](
       }
     }
 
-  // 64-bit unsigned multiplication was adopted from the great Hacker's Delight function
-  // (Henry S. Warren, Hacker's Delight, Addison-Wesley, 2nd edition, Fig. 8.2)
-  // https://doc.lagout.org/security/Hackers%20Delight.pdf
-  private[this] def multiplyHigh(x: Long, y: Long): Long = {
-    val xl = x & 0xFFFFFFFFL
-    val xh = x >>> 32
-    val yl = y & 0xFFFFFFFFL
-    val yh = y >>> 32
-    val t = xh * yl + (xl * yl >>> 32)
-    xh * yh + (t >>> 32) + (xl * yh + (t & 0xFFFFFFFFL) >>> 32)
-  }
+  private[this] def unsignedMultiplyHigh(x: Long, y: Long): Long =
+    Math.multiplyHigh(x, y) + ((x >> 63) & y) + ((y >> 63) & x)
 
   private[this] def addExp(e2: Int, e10: Int): Int =
     (e10 * 108853 >> 15) + e2 + 1 // (e10 * Math.log(10) / Math.log(2)).toInt + e2 + 1