Performance improvements on JDK 17 are too small to drop JDK 8 support now.
## Before
```
[info] Benchmark                                    (size)   Mode  Cnt       Score       Error  Units
[info] ArrayOfDoublesReading.jsoniterScala             128  thrpt    5  226446.023 ± 49427.780  ops/s
[info] ArrayOfDoublesWriting.jsoniterScala             128  thrpt    5  168589.640 ± 11525.663  ops/s
[info] ArrayOfDoublesWriting.jsoniterScalaPrealloc     128  thrpt    5  160991.678 ± 33580.137  ops/s
[info] ArrayOfFloatsReading.jsoniterScala              128  thrpt    5  322841.853 ±  3838.026  ops/s
[info] ArrayOfFloatsWriting.jsoniterScala              128  thrpt    5  273020.114 ±  6094.720  ops/s
[info] ArrayOfFloatsWriting.jsoniterScalaPrealloc      128  thrpt    5  275820.840 ± 28968.697  ops/s
```
## After
```
[info] Benchmark                                    (size)   Mode  Cnt       Score       Error  Units
[info] ArrayOfDoublesReading.jsoniterScala             128  thrpt    5  231053.087 ± 36165.692  ops/s
[info] ArrayOfDoublesWriting.jsoniterScala             128  thrpt    5  173660.841 ± 20153.898  ops/s
[info] ArrayOfDoublesWriting.jsoniterScalaPrealloc     128  thrpt    5  180128.598 ± 42129.485  ops/s
[info] ArrayOfFloatsReading.jsoniterScala              128  thrpt    5  314440.050 ± 29804.469  ops/s
[info] ArrayOfFloatsWriting.jsoniterScala              128  thrpt    5  272804.634 ±  3313.528  ops/s
[info] ArrayOfFloatsWriting.jsoniterScalaPrealloc      128  thrpt    5  289814.814 ±  1391.401  ops/s
```